chore: move user logic from 'fetchdata/api/' to 'users/api/'
