+++ fetchdata/api/serializers.py
@@ -1,5 +1,4 @@
 from rest_framework import serializers
-from django.contrib.auth.models import User
 
 from core.models import Place
 
@@ -28,16 +27,3 @@ class PlaceSerializer(serializers.ModelSerializer):
     class Meta:
         model = Place
         fields = "__all__"
-
-
-class UserSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = User
-        fields = ["username", "password"]
-        extra_kwargs = {"password": {"write_only": True}}
-
-    def create(self, validated_data):
-        username = validated_data.get("username")
-        password = validated_data.get("password")
-        user = User.objects.create_user(username=username, password=password)
-        return user
